requirements.txt
pip~=21.2.4
wheel~=0.37.0
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
